pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["RuslanUC <dev_ruslan_uc@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
tortoise-orm = {extras = ["mysql"], version = "^0.20.0"}
pycryptodome = "^3.20.0"
bcrypt = "^4.1.2"
python-magic = "^0.4.27"
pillow = "^10.2.0"
s3lite = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
